+++ src/Collections/CollectionMacrosServiceProvider.php
@@ -68,10 +68,16 @@ class CollectionMacrosServiceProvider extends ServiceProvider
         });
 
         /**
-         * Transpose (flip) a collection matrix (array of arrays) while keeping
-         * its columns and row headers intact.
+         * Transpose (flip) a collection matrix (array of arrays) while keeping its columns and row headers intact.
+         *
+         * Please note that a row missing a column another row does have can only occur for one column. It cannot
+         * parse more than one missing column.
          */
-        Collection::macro('transposeWithKeys', function (array $rows) {
+        Collection::macro('transposeWithKeys', function (array $rows = null) {
+            $rows = $rows ?? $this->values()->reduce(function (array $rows, array $values) {
+                    return array_unique(array_merge($rows, array_keys($values)));
+                }, []);
+
             $keys = $this->keys()->toArray();
 
             // Transpose the matrix
@@ -80,7 +86,7 @@ class CollectionMacrosServiceProvider extends ServiceProvider
                 return array_combine($keys, $items);
             }, ...$this->values());
 
-            // Add the row headers
+            // Add the new row headers
             $items = array_combine($rows, $items);
 
             return new static($items);