+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Changed
 
 - Locked down dependencies more strictly, but allow optional use of each helper
+- Added type hints where possible
 - Tweaked Travis test script
 - Renamed global method helpers service provider
 - Renamed collection service provider
@@ -33,7 +34,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Extracted base Eloquent model (moved to [laravel-resource-flow](https://github.com/sebastiaanluca/laravel-resource-flow))
 - Extracted queueable job (moved to [laravel-resource-flow](https://github.com/sebastiaanluca/laravel-resource-flow))
 - Removed `mapWithIntegerKeys` collection macro (fixed in Laravel 5.4.x)
-- Removed HTML and form helpers
+- Removed HTML and form helpers (tip: use Spatie's macroable https://github.com/spatie/laravel-html package instead)
 
 ## 0.5.2 (2017-02-14)